removed env from save\

+++ module/ActorCritic/models/ANN.py
@@ -17,9 +17,8 @@ class ANN_V1(nn.Module):
         self.critic_fc2 = nn.Linear(256, 1)
         
     def forward(self, x):
-        
-        state = x.to(torch.float32)
-        
+        state = x
+
         probs = F.relu(self.actor_fc1(state))
         probs = self.head(self.actor_fc2(probs))
